+++ jeap-process-context-domain/src/main/java/ch/admin/bit/jeap/processcontext/domain/maintenance/MaintenanceTaskState.java
@@ -3,7 +3,6 @@ package ch.admin.bit.jeap.processcontext.domain.maintenance;
 public enum MaintenanceTaskState {
     CREATED,
     EVENT_QUEUED,
-    PROCESSING,
     SUCCEEDED,
     NOT_FOUND,
     FAILED;